requirements.txt
future
emoji
beautifulsoup4
requests
sqlalchemy
python-telegram-bot==13.3
psycopg2-binary
feedparser
markdown2
spongemock
tswift
zalgo-text
geopy
nltk
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
gpytranslate
speedtest-cli
spamwatch
telethon==1.16.4
configparser
wikipedia
pyrogram
tgcrypto
alphabet_detector
pyrate-limiter
bleach
cachetools
loguru
PyYAML>=5.1.2
Intellivoid-SPB
rich
psutil
